Add files via upload
So this is the dashboard of recruiter CRM. Under candidate and job profile section, you can see the, all the candidate names.
We will click on the candidate, uhm, and it will show us all the details of that candidate. And on the top right section, we can see add call logs, add tags, and admitting.
We'll click on add tags and give it a name like, uh, schedule interview. You can set up a due reminder and the starting time also and what type it is.
It's like interview scheduling and we'll click on add. Once it's added, it will show you, it will appear on the tags.
It's like interview scheduling and we'll click on add. Once it's added, it will show you, it will appear on the tags.